--- src/gui/static/src/app/components/pages/wallets/address-detail/wallet-detail.component.html
@@ -1,11 +1,0 @@
-<md-list>
-  <h3 md-subheader>Addresses</h3>
-  <md-list-item *ngFor="let address of wallet.entries">
-    <i md-list-icon (click)="showQr(address)" class="fa fa-qrcode"></i>
-    <h4 md-line><span class="click-to-copy" (click)="copyAddress(address.address)" (mouseenter)="toggleClass()" (mouseleave)="toggleClass()">{{address.address}} <span data-label="Copied" class="copy-label hidden">Copy</span></span> - {{ address.balance | sky }} ({{ address.hours ? address.hours : 0 }} Coin Hours)</h4>
-  </md-list-item>
-  <div class="button-line">
-    <a md-raised-button color="primary" (click)="renameWallet()">Rename wallet</a>
-    <a md-raised-button color="primary" (click)="addAddress()">Add address</a>
-  </div>
-</md-list>
