+++ src/gui/static/src/app/components/pages/wallets/wallet-detail/wallet-detail.component.html
@@ -1,9 +1,9 @@
 <ng-container *ngFor="let address of (wallet ? wallet.addresses : []); let i = index">
   <div class="-record" *ngIf="!wallet.hideEmpty || address.coins">
     <div class="id-column">{{ i + 1 }}</div>
-    <div class="address-column">
+    <div class="address-column" (click)="copyAddress(address.address, i)" (mouseenter)="toggleClass(i)" (mouseleave)="toggleClass(i)">
       <img src="../../../../../assets/img/qr-code-black.png">
-      <p>{{ address.address }}</p>
+      <p class="click-to-copy">{{ address.address }} <span data-label="Copied" class="copy-label hidden">Copy</span></p>
     </div>
     <div class="hours-column">{{ address.hours ? address.hours : 0 }}</div>
     <div class="coins-column">{{ address.coins ? address.coins : 0 }}</div>
